Many changes
In order for native compilation to work well, byte compilation has to
too. I'm attempting to reorganize doom core to fix warnings and make
compilation the feasible default. The way dependencies are organized is
extremely delicate and the tiniest change can break Doom entirely. In my
experience so far, it's been a tossup whether any given change to doom
upstream will or will not break byte compilation. Doom's solution to
this is to just suppress all unsavory warnings, both in flycheck and
'bin/doom compile'.

Also, some of these changes are premature optimization or aesthetic
preference.